porting to Desktop Windows at Carolina Kiyoko Dell at ckribeiro LAN at Canal 2 - Avenida Bernardino de Campos Santos SP, Brazil...
at machine Dell at network NetVirtua 10Mbps at Santos, SP - Brazil, at laptop Dell XPS L502X processor i7 2670QM 8GB RAM 1 TB Disk 1.5TB Backup Samsung Disk at Santos SP, Brazil

+++ appseed/ca/userpresence/userpresence_user_presence.cpp
@@ -23,7 +23,7 @@ namespace userpresence
    void presence::defer_pulse_user_presence()
    {
 
-      int iStatus = m_iShortStatusWayTag;
+      int iStatus = (int) m_iShortStatusWayTag;
       if(::get_tick_count() - m_dwLastActivity < ((1984 + 1977) * 4))
       {
          iStatus = status_online;
